move dtype to tensor ops
