+++ ttnn/core/tensor/tensor.cpp
@@ -692,4 +692,8 @@ const std::optional<NdShardSpec>& Tensor::nd_shard_spec() const { return this->m
 
 const TensorTopology& Tensor::tensor_topology() const { return this->tensor_attributes->get_tensor_topology(); }
 
+namespace operations {
+Tensor to_dtype(const Tensor& tensor, DataType dtype) { return tensor_ops::tensor_to_dtype(tensor, dtype); }
+}  // namespace operations
+
 }  // namespace tt::tt_metal
